requirements.txt
python-telegram-bot
watchdog
git+https://github.com/qdot/np-telegram-bot.git
